+++ packages/react-core/src/types.d.ts
@@ -1,9 +1,6 @@
 import { TILE_FORMATS } from '@deck.gl/carto'
 import { AggregationTypes } from './operations/aggregation/AggregationTypes';
 import { Geometry } from 'geojson';
-import {
-  FilterTypes
-} from './filters/FilterQueryBuilder';
 
 export type AggregationFunctions = {
   [AggregationTypes.COUNT]: Function,
@@ -31,5 +28,3 @@ export type TileFeatures = {
 }
 
 export type TileFeaturesResponse = Record<string, unknown>[] | []
-
-export type _FilterTypes = typeof FilterTypes;